Convert Dropdown.test to TS; update snapshots. All tests pass
